+++ checkpoint/orbax/checkpoint/experimental/v1/_src/handlers/generic_handler.py
@@ -1,0 +1,71 @@
+# Copyright 2024 The Orbax Authors.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+"""GenericHandler class."""
+
+from typing import Any, Awaitable, Generic
+from orbax.checkpoint.experimental.v1._src.handlers import types as handler_types
+from orbax.checkpoint.experimental.v1._src.path import types as path_types
+
+
+class GenericHandler(
+    handler_types.CheckpointableHandler[handler_types.T, handler_types.T],
+    Generic[handler_types.T],
+):
+  """Serializes/deserializes a Checkpointable."""
+
+  async def save(
+      self,
+      directory: path_types.PathAwaitingCreation,
+      checkpointable: handler_types.StatefulCheckpointable[handler_types.T],
+  ) -> Awaitable[None]:
+    return await checkpointable.save(directory)
+
+  async def _background_load(
+      self,
+      directory: path_types.Path,
+      abstract_checkpointable: handler_types.StatefulCheckpointable[
+          handler_types.T
+      ],
+  ) -> handler_types.T:
+    await abstract_checkpointable.load(directory)
+    return abstract_checkpointable
+
+  async def load(
+      self,
+      directory: path_types.Path,
+      abstract_checkpointable: (
+          handler_types.StatefulCheckpointable[handler_types.T] | None
+      ) = None,
+  ) -> Awaitable[handler_types.T]:
+    if abstract_checkpointable is None:
+      raise ValueError(
+          'Abstract checkpointable is required for GenericHandler.load.'
+      )
+    return self._background_load(directory, abstract_checkpointable)
+
+  async def metadata(self, directory: path_types.Path) -> handler_types.T:
+    # TODO(yaning): Implement Metadata.
+    raise NotImplementedError()
+
+  def is_handleable(self, checkpointable: Any) -> bool:
+    # TODO(yaning): Add test for a class that partially implements
+    # StatefulCheckpointable, and a test that inherits from
+    # StatefulCheckpointable.
+    return isinstance(checkpointable, handler_types.StatefulCheckpointable)
+
+  def is_abstract_handleable(self, abstract_checkpointable: Any) -> bool | None:
+    return isinstance(
+        abstract_checkpointable, handler_types.StatefulCheckpointable
+    )