+++ checkpoint/CHANGELOG.md
@@ -25,6 +25,7 @@ would want them to be preferred.
 - `PartsOf` structure which holds a PyTree whose leaf nodes may be missing.
 - #v1 Add compatibility tests for save-by-v0-load-by-v1 and also fix code.
 - #v1 Enable v1 users to load v0 checkpoints without using contexts.
+- #v1 Add `StatefulCheckpointable` and `GenericHandler`.
 
 ## [0.11.12] - 2025-04-09